PMM-15249 Reuse the gather helpers across the metrics tests.
